This makes it so we can actually be in development mode.
I copied the code for sprockets from a sinatra recipe, but when I set
the environment to that it means our app won't run in development mode
by default, which means we don't get nice stack traces or anything like
that. It also means the auto reload plugin doesn't work, and now it
should.

Turns out you can just call it sprockets and change the references
elsewhere and everything works fine.