Render date and time fields when field is a datetime.
This is determined by it being called "datetime", by it ending in
"_datetime", or just being a Time object when deserialized.

I pulled in active_support to use the "present?" method, since it is
already installed via another gem. I'm not totally sold on it though, I
would be ok to replace it with something else.