Add the ability to use a multiple select box.
If the YML key is the same name as a key from the .tint.yml options
hash, we will display a multiselect. This also takes precedence over
normal array input displays.

There is a hidden input to submit an array, because if you select
nothing in the multiselect that will not submit to the server, which
results in the key being deleted. The hidden input fixes this behaviour.
Because the hidden input submits an array with one entry that has a
value of "" we now reject blank values in our process_form_data method.