Refactor the global constants to use Tint::Site
Routes are all relative to /:site now and everything loads the site at
that id to determine its settings.  If SITE_PATH is set, then no DB is
created and instead a Tint::Site with the correct id is ginned up.
Everything should work as it did before so long as SITE_PATH is set.

This basically gets us to multitenancy.  Though users cannot yet create
a site (#42) and we do not yet clone or sync with a remote to
meaningfully create the local cache.

However, this *does* contain enough of the permissions bits to restrict
a user to only see their own sites.

Closes #41