Change to have a user model with methods on it.
Since I changed commit_with to have keyword arguments, passing it the
user hash caused it to think I was passing keyword arguments that did
not exist. This is apparently the desired behaviour:
https://bugs.ruby-lang.org/issues/11967

Passing a user object instead of a hash fixes it, since it will not
think that I am passing multiple keyword arguments.