Raise mass threshold
Codeclimate reports spec-style tests as identical too often because they
share the describe/it/call method nested structure.  It looks like
something that can be refactored, but really it's just a syncatic hack
that spec-style tests use.

Using a different mass only in tests using such a DSL would be nice, but
this is easiest for now.