Fix frontmatter filename/root array element issue.
The problem was that the frontmatter from the filename would always
return an empty hash. However, if the root element of a YML document (or
frontmatter) was an array, we would attempt to merge an array into a
hash, which does not work.

Our solution is to check if there is any frontmatter from the filename.
If not, then just return the frontmatter from the content. If there is
frontmatter in the filename and the root element of the frontmatter is
not a Hash, then we raise an IncompatibleFrontmatter exception.

A future commit will handle this exception in order to display a proper
error page to the user.