Basic New Site flow
Instead of listing every possible option for creation right on the
homepage, break it up a bit.

Now the homepage is just existant sites, and an "Add Site" button.

This takes them to a page listing configured git providers, with an "Add
Provider" button and also a manual site creation form at the very
bottom.

Selecting a configured git provider (you can have more than one github
configured, so disambiguation is done with username) shows the same list
of repos (minus ones already used as sites) as before, with an "Add"
button for each one.

Closes #324