Reorganize editor views
There are conceputally four editor views:

1. WYSIWYG editor with frontmatter editor
2. Frontmatter-only editor (content editor hidden)
3. Source editor (default on non-WYSIWYG files with no frontmatter)
4. Binary file editor (with optional filename frontmatter)

"Page title" is pulled out as special only if the key is present and not
editing a YAML file.

Hiding title and content editor on YAML files makes the page a data-only
editor, so we don't need a seperate view for that.  1 and 2 are thus
both a single view file now.

File with no frontmatter at all go straight to the source editor, since
they would otherwise have a hidden content portion (not WYSIWYG) and no
data fields... which is just an empty page.

Closes #314
Closes #316