Make sure image files never think they have frontmatter.
Even if there are frontmatter filename rules in the site config.
Currently I cannot think of any reason why it would make sense to have
frontmatter in image filenames. If that proves to be desired, then
removing this code would make sense.

However, this is to fix a specific bug, and that is when you are
visiting the page for an image file, it was (before this commit) showing
you the editor page, with nothing on it and a link to source. Now that
this commit makes it so the `frontmatter?` method always returns false
for images, it falls through in the controller to the case where the
binary file renderer takes over. If the code is changed so that
frontmatter can exist for images, then this bug will need to be fixed in
another way.

The extra requires have been added because when running the test files
individually or when opening up IRB and requiring a file, these requires
needed to be added.